refactor: add Board model to GameController
