+++ jogo-oito/f:w
@@ -1,0 +1,12 @@
+export default class Board {
+  private matrix: number[][]
+
+  constructor(matrix: number[][]) {
+    this.matrix = matrix
+  }
+
+  getMatrixAsArray() {
+    return this.matrix.flat()
+  }
+
+}
